Define parameter interpreters inline
