+++ lib/exoskeleton/src/args/soundness+exoskeleton-args.scala
@@ -34,8 +34,7 @@ package soundness
 
 export exoskeleton
 . { Argument, Arguments, Cli, CliInterpreter, Flag, Interpretable, FlagParameters,
-    PosixCliInterpreter, PosixParameters, Shell, SimpleParameterInterpreter, Subcommand, Suggestion,
-    Discoverable, Suggestible, Switch, arguments }
+    PosixParameters, Shell, Subcommand, Suggestion, Discoverable, Suggestible, Switch, arguments }
 
 package parameterInterpretation:
   export exoskeleton.parameterInterpretation.{simple, posix}
